進捗管理についてバックエンドで整合性をとりつつ実装 #4

+++ src/main/kotlin/com/j15/backend/presentation/dto/response/UserProgressResponse.kt
@@ -6,7 +6,7 @@ import java.time.Instant
 // 進捗状態レスポンスDTO
 data class UserProgressResponse(
         val userId: String,
-        val progressPercentage: Double,
+        val progressPercentage: Int,
         val clearedCount: Int,
         val remainingCount: Int,
         val totalSections: Int,
